backend/requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
wonderwords
langchain
openai
wandb
textstat
pillow
rich
pandas
spacy